Example extent

The example datasets contain the sf objects ext and ext_eur that can be used to call basemap and the associated functions.

Usage

data(ext)

data(ext_eur)

Format

sf object

An object of class sf (inherits from data.frame) with 1 rows and 3 columns.