Set, get and reset basemaps defaults

These functions set, get or reset the defaults of all map arguments passed to basemap and associated functions.

set_defaults(
  ext = NULL,
  map_service = NULL,
  map_type = NULL,
  map_res = NULL,
  map_token = NULL,
  map_dir = NULL
)

get_defaults()

reset_defaults()

Arguments

ext
extent to be covered by the basemap as any spatial class supported by st_bbox.

map_service
character, a map service, either "osm", "carto" or "mapbox". Default is "osm".

map_type
character, a map type, e.g. "streets". For a full list of available map types, see get_maptypes.

map_res
numeric, resolution of base map in range from 0 to 1.

map_token
character, mapbox authentification token for mapbox basemaps. Register at https://www.mapbox.com/ to get a mapbox token. Mapbox is free of charge after registration for up to 50.000 map requests per month. Ignored, if map_service = "osm".

map_dir
character, cache directory where downloaded basemap tiles will be stored. By default, a temporary directory is used, which is destroyed when the session is terminated.

Value

For get_defaults, a list of defaults, otherwise none.

Examples

library(basemaps)
data(ext)

# set defaults for the basemap
set_defaults(ext = ext, map_service = "osm", map_type = "terrain_bg")
# for mapbox maps, you need a map_token. Register for free at mapbox.com to get a token

# get defaults
get_defaults()

#> $map_service
#> [1] "osm"
#> 
#> $map_type
#> [1] "terrain_bg"
#> 
#> $map_res
#> [1] 1
#> 
#> $map_token
#> [1] NA
#> 
#> $map_dir
#> [1] "/tmp/RtmpU8knk1/basemaps/"
#> 
#> $ext
#> Simple feature collection with 1 feature and 2 fields
#> geometry type:  POLYGON
#> dimension:      XY
#> bbox:           xmin: 1225767 ymin: 6034383 xmax: 1237651 ymax: 6045498
#> projected CRS:  WGS 84 / Pseudo-Mercator
#>   X_leaflet_id feature_type                       geometry
#> 1          446    rectangle POLYGON ((1225767 6034383, ...
#> 

if (FALSE) {
# load and return basemap map as raster (default)
map <- basemap()
}

# reset defaults
reset_defaults()