Get all supported map types

This function returns every supported map type that can be used as input to the map_type argument of set_defaults, basemap or associated functions.

get_maptypes(map_service = NULL)

Arguments

map_service
character, optional, either "osm", "osm_stamen", "osm_thunderforest", "carto", "mapbox" or "esri". Otherwise, a list of map types for both services is returned.

Source

"osm": Open Street Map contributors (https://www.openstreetmap.org/copyright), Open Topo Map (https://opentopomap.org/), Martin Tesar (http://mtbmap.cz/)

"osm_stamen": Stamen (http://maps.stamen.com/), Open Street Map contributors (https://www.openstreetmap.org/copyright)

"osm_thunderforest": Thunderforest (https://www.thunderforest.com/), Open Street Map contributors (https://www.openstreetmap.org/copyright)

"carto": Carto (https://carto.com/)

"mapbox": Mapbox (https://www.mapbox.com)

"esri": Esri (https://www.esri.com/)

Value

A character vector of supported map types

See also

Examples

# for all services
get_maptypes()

#> $osm
#>  [1] "streets"      "streets_de"   "streets_fr"   "humanitarian" "topographic" 
#>  [6] "hike"         "hillshade"    "grayscale"    "no_labels"    "mtb"         
#> 
#> $osm_stamen
#> [1] "toner"      "toner_bg"   "toner_lite" "terrain"    "terrain_bg"
#> [6] "watercolor"
#> 
#> $osm_thunderforest
#>  [1] "cycle"          "transport"      "landscape"      "outdoors"      
#>  [5] "transport_dark" "spinal"         "pioneer"        "mobile_atlas"  
#>  [9] "neighbourhood"  "atlas"         
#> 
#> $carto
#>  [1] "light"                "light_no_labels"      "light_only_labels"   
#>  [4] "dark"                 "dark_no_labels"       "dark_only_labels"    
#>  [7] "voyager"              "voyager_no_labels"    "voyager_only_labels" 
#> [10] "voyager_labels_under"
#> 
#> $mapbox
#> [1] "streets"   "outdoors"  "light"     "dark"      "satellite" "hybrid"   
#> [7] "terrain"  
#> 
#> $esri
#>  [1] "natgeo_world_map"                     
#>  [2] "usa_topo_maps"                        
#>  [3] "world_imagery"                        
#>  [4] "world_physical_map"                   
#>  [5] "world_shaded_relief"                  
#>  [6] "world_street_map"                     
#>  [7] "world_terrain_base"                   
#>  [8] "world_topo_map"                       
#>  [9] "world_dark_gray_base"                 
#> [10] "world_dark_gray_reference"            
#> [11] "world_light_gray_base"                
#> [12] "world_light_gray_reference"           
#> [13] "world_hillshade_dark"                 
#> [14] "world_hillshade"                      
#> [15] "world_ocean_base"                     
#> [16] "world_ocean_reference"                
#> [17] "antarctic_imagery"                    
#> [18] "arctic_imagery"                       
#> [19] "arctic_ocean_base"                    
#> [20] "arctic_ocean_reference"               
#> [21] "world_boundaries_and_places_alternate"
#> [22] "world_boundaries_and_places"          
#> [23] "world_reference_overlay"              
#> [24] "world_transportation"                 
#> [25] "delorme_world_base_map"               
#> [26] "world_navigation_charts"              
#> 

# for osm only
get_maptypes("osm")

#>  [1] "streets"      "streets_de"   "streets_fr"   "humanitarian" "topographic" 
#>  [6] "hike"         "hillshade"    "grayscale"    "no_labels"    "mtb"         

# or
get_maptypes()$osm

#>  [1] "streets"      "streets_de"   "streets_fr"   "humanitarian" "topographic" 
#>  [6] "hike"         "hillshade"    "grayscale"    "no_labels"    "mtb"         

# for mapbox only
get_maptypes("mapbox")

#> [1] "streets"   "outdoors"  "light"     "dark"      "satellite" "hybrid"   
#> [7] "terrain"  

# or
get_maptypes()$mapbox

#> [1] "streets"   "outdoors"  "light"     "dark"      "satellite" "hybrid"   
#> [7] "terrain"  

# same for all other map services